dictionary: added parse
